gfg-scraper/requirements.txt
streamlit
requests
beautifulsoup4
lxml
openai
langchain
langchain-openai
python-dotenv
